refactor: Remove authentication use case and adopt test factories
Removes the `AuthenticateSalespersonUseCase` as it is no longer required.

To improve the testing suite, this change adds the Faker.js library and introduces a `makeSalesperson` factory. This simplifies the creation of test entities and makes tests cleaner and more robust. The salesperson registration test is updated to use this new factory.

Additionally, the `FakeHasher` test double is enhanced to implement hash comparison, making it more versatile for testing cryptographic operations.